Move PBftChainState to its own module
